Fix internal structures data types, add withdraw handler (implementation in storage is not completed)

+++ internal/client/accrual.go
@@ -16,9 +16,9 @@ type (
 	}
 
 	accrualStatus struct {
-		Order   string `json:"order"`
-		Status  string `json:"status"`
-		Accrual int    `json:"accrual"`
+		Order   string  `json:"order"`
+		Status  string  `json:"status"`
+		Accrual float32 `json:"accrual"`
 	}
 )
